Centralize settings registration in Options class
Move register_settings from WP_Admin\Settings to Options class for
better organization and to support both admin and REST API contexts.

- Add register_settings() method to Options class
- Hook registration to both admin_init and rest_api_init
- Remove duplicate registration from WP_Admin\Settings
- Clean up unused imports in Settings class
- Remove translation wrappers from non-user-facing descriptions
