+++ includes/wp-admin/class-settings.php
@@ -8,8 +8,6 @@
 namespace Activitypub\WP_Admin;
 
 use Activitypub\Collection\Actors;
-use Activitypub\Model\Blog;
-use Activitypub\Sanitize;
 
 use function Activitypub\user_can_activitypub;
 
@@ -21,364 +19,9 @@ class Settings {
 	 * Initialize the class, registering WordPress hooks,
 	 */
 	public static function init() {
-		\add_action( 'admin_init', array( self::class, 'register_settings' ), 11 );
 		\add_action( 'admin_menu', array( self::class, 'add_settings_page' ) );
 
 		\add_action( 'load-settings_page_activitypub', array( self::class, 'handle_welcome_query_arg' ) );
-	}
-
-	/**
-	 * Register ActivityPub settings
-	 */
-	public static function register_settings() {
-		\register_setting(
-			'activitypub',
-			'activitypub_post_content_type',
-			array(
-				'type'         => 'string',
-				'description'  => \__( 'Use title and link, summary, full or custom content', 'activitypub' ),
-				'show_in_rest' => array(
-					'schema' => array(
-						'enum' => array( 'title', 'excerpt', 'content' ),
-					),
-				),
-				'default'      => 'content',
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_custom_post_content',
-			array(
-				'type'         => 'string',
-				'description'  => \__( 'Define your own custom post template', 'activitypub' ),
-				'show_in_rest' => true,
-				'default'      => ACTIVITYPUB_CUSTOM_POST_CONTENT,
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_max_image_attachments',
-			array(
-				'type'              => 'integer',
-				'description'       => \__( 'Number of images to attach to posts.', 'activitypub' ),
-				'default'           => ACTIVITYPUB_MAX_IMAGE_ATTACHMENTS,
-				'sanitize_callback' => function ( $value ) {
-					return \is_numeric( $value ) ? \absint( $value ) : ACTIVITYPUB_MAX_IMAGE_ATTACHMENTS;
-				},
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_use_hashtags',
-			array(
-				'type'        => 'boolean',
-				'description' => \__( 'Add hashtags in the content as native tags and replace the #tag with the tag-link', 'activitypub' ),
-				'default'     => '0',
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_use_opengraph',
-			array(
-				'type'        => 'boolean',
-				'description' => \__( 'Automatically add "fediverse:creator" OpenGraph tags for Authors and the Blog-User.', 'activitypub' ),
-				'default'     => '1',
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_support_post_types',
-			array(
-				'type'         => 'string',
-				'description'  => \esc_html__( 'Enable ActivityPub support for post types', 'activitypub' ),
-				'show_in_rest' => true,
-				'default'      => array( 'post' ),
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_actor_mode',
-			array(
-				'type'        => 'integer',
-				'description' => \__( 'Choose your preferred Actor-Mode.', 'activitypub' ),
-				'default'     => ACTIVITYPUB_ACTOR_MODE,
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_attribution_domains',
-			array(
-				'type'              => 'string',
-				'description'       => \__( 'Websites allowed to credit you.', 'activitypub' ),
-				'default'           => \Activitypub\home_host(),
-				'sanitize_callback' => array( Sanitize::class, 'host_list' ),
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_allow_likes',
-			array(
-				'type'              => 'integer',
-				'description'       => \__( 'Allow likes.', 'activitypub' ),
-				'default'           => '1',
-				'sanitize_callback' => 'absint',
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_allow_reposts',
-			array(
-				'type'              => 'integer',
-				'description'       => \__( 'Allow reposts.', 'activitypub' ),
-				'default'           => '1',
-				'sanitize_callback' => 'absint',
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_auto_approve_reactions',
-			array(
-				'type'              => 'integer',
-				'description'       => \__( 'Auto approve Reactions.', 'activitypub' ),
-				'default'           => '0',
-				'sanitize_callback' => 'absint',
-			)
-		);
-
-		\register_setting(
-			'activitypub',
-			'activitypub_relays',
-			array(
-				'type'              => 'array',
-				'description'       => \__( 'Relays', 'activitypub' ),
-				'default'           => array(),
-				'sanitize_callback' => array( Sanitize::class, 'url_list' ),
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_outbox_purge_days',
-			array(
-				'type'        => 'integer',
-				'description' => \__( 'Number of days to keep items in the Outbox.', 'activitypub' ),
-				'default'     => 180,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_inbox_purge_days',
-			array(
-				'type'        => 'integer',
-				'description' => \__( 'Number of days to keep items in the Inbox.', 'activitypub' ),
-				'default'     => 180,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_ap_post_purge_days',
-			array(
-				'type'        => 'integer',
-				'description' => 'Number of days to keep remote posts.',
-				'default'     => 30,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_vary_header',
-			array(
-				'type'        => 'boolean',
-				'description' => \__( 'Add the Vary header to the ActivityPub response.', 'activitypub' ),
-				'default'     => true,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_content_negotiation',
-			array(
-				'type'        => 'boolean',
-				'description' => 'Enable content negotiation.',
-				'default'     => true,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_authorized_fetch',
-			array(
-				'type'        => 'boolean',
-				'description' => \__( 'Require HTTP signature authentication.', 'activitypub' ),
-				'default'     => false,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_rfc9421_signature',
-			array(
-				'type'        => 'boolean',
-				'description' => 'Use RFC-9421 signature.',
-				'default'     => false,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_following_ui',
-			array(
-				'type'        => 'boolean',
-				'description' => 'Show Following UI in admin menus and settings.',
-				'default'     => false,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_create_posts',
-			array(
-				'type'        => 'boolean',
-				'description' => 'Allow creating posts via ActivityPub.',
-				'default'     => false,
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_object_type',
-			array(
-				'type'         => 'string',
-				'description'  => \__( 'The Activity-Object-Type', 'activitypub' ),
-				'show_in_rest' => array(
-					'schema' => array(
-						'enum' => array( 'note', 'wordpress-post-format' ),
-					),
-				),
-				'default'      => ACTIVITYPUB_DEFAULT_OBJECT_TYPE,
-			)
-		);
-
-		// Blog-User Settings.
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_blog_description',
-			array(
-				'type'         => 'string',
-				'description'  => \esc_html__( 'The Description of the Blog-User', 'activitypub' ),
-				'show_in_rest' => true,
-				'default'      => '',
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_blog_identifier',
-			array(
-				'type'              => 'string',
-				'description'       => \esc_html__( 'The Identifier of the Blog-User', 'activitypub' ),
-				'show_in_rest'      => true,
-				'default'           => Blog::get_default_username(),
-				'sanitize_callback' => array( Sanitize::class, 'blog_identifier' ),
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_header_image',
-			array(
-				'type'        => 'integer',
-				'description' => \__( 'The Attachment-ID of the Sites Header-Image', 'activitypub' ),
-				'default'     => null,
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_blog_user_mailer_new_dm',
-			array(
-				'type'        => 'integer',
-				'description' => 'Send a notification when someone sends a user of the blog a direct message.',
-				'default'     => 1,
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_blog_user_mailer_new_follower',
-			array(
-				'type'        => 'integer',
-				'description' => 'Send a notification when someone starts to follow a user of the blog.',
-				'default'     => 1,
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_blog_user_mailer_new_mention',
-			array(
-				'type'        => 'integer',
-				'description' => 'Send a notification when someone mentions a user of the blog.',
-				'default'     => 1,
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_blog_user_also_known_as',
-			array(
-				'type'              => 'array',
-				'description'       => 'An array of URLs that the blog user is known by.',
-				'default'           => array(),
-				'sanitize_callback' => array( Sanitize::class, 'identifier_list' ),
-			)
-		);
-
-		\register_setting(
-			'activitypub_blog',
-			'activitypub_hide_social_graph',
-			array(
-				'type'              => 'integer',
-				'description'       => 'Hide Followers and Followings on Profile.',
-				'default'           => 0,
-				'sanitize_callback' => 'absint',
-			)
-		);
-
-		// Moderation settings.
-		\register_setting(
-			'activitypub',
-			'activitypub_site_blocked_actors',
-			array(
-				'type'              => 'array',
-				'description'       => 'Site-wide blocked ActivityPub actors.',
-				'default'           => array(),
-				'sanitize_callback' => array( Sanitize::class, 'identifier_list' ),
-			)
-		);
-
-		\register_setting(
-			'activitypub_advanced',
-			'activitypub_relay_mode',
-			array(
-				'type'              => 'integer',
-				'description'       => 'Enable relay mode to forward public activities to all followers.',
-				'default'           => 0,
-				'sanitize_callback' => 'absint',
-			)
-		);
 	}
 
 	/**
